+++ src/xrGame/ai/monsters/anomaly_detector.cpp
@@ -36,8 +36,8 @@ void CAnomalyDetector::update_schedule()
 	if (m_storage.empty())
 		return;
 
-	xr_vector<u16> temp_out_restrictors;
-	xr_vector<u16> temp_in_restrictors;
+	xr_vector<u32> temp_out_restrictors;
+	xr_vector<u32> temp_in_restrictors;
 
 	temp_in_restrictors.reserve(m_storage.size());
