+++ src/xrGame/xrServer_CL_connect.cpp
@@ -9,7 +9,7 @@
 void xrServer::Perform_connect_spawn(CSE_Abstract* E, xrClientData* CL, NET_Packet& P)
 {
 	P.B.count = 0;
-	xr_vector<u16>::iterator it = std::find(conn_spawned_ids.begin(), conn_spawned_ids.end(), E->ID);
+	xr_vector<u32>::iterator it = std::find(conn_spawned_ids.begin(), conn_spawned_ids.end(), E->ID);
 	if (it != conn_spawned_ids.end())
 	{
 		//.		Msg("Rejecting redundant SPAWN data [%d]", E->ID);
@@ -87,7 +87,7 @@ void xrServer::SendConnectionData(IClient* _CL)
 
 	/*
 		Msg("--- Our sended SPAWN IDs:");
-		xr_vector<u16>::iterator it = conn_spawned_ids.begin();
+		xr_vector<u32>::iterator it = conn_spawned_ids.begin();
 		for (; it != conn_spawned_ids.end(); ++it)
 		{
 			Msg("%d", *it);
